refactor: move functions to service file
